Now with findNextSquare
Updated BoardSquares with findNextSquare. Feed findNextSquare the
current x, y board square values of an enemy, and it will return the
INDEX in boardSquaresArray for the next square they need to go to. Could
quickly be modified to return the next x, y instead. Also, MainDemo
updated with example call.